% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrthoDotPlot.R
\name{OrthoDotPlot}
\alias{OrthoDotPlot}
\alias{SplitDotPlotGG}
\title{Orthogonal DotPlot}
\usage{
OrthoDotPlot(
  object,
  features,
  assay = NULL,
  cols = c("lightgrey", "blue"),
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  idents = NULL,
  cluster.idents = FALSE,
  scale = TRUE,
  legend = NULL,
  scale.by = "radius",
  keep.scale = "feature",
  scale.min = NA,
  scale.max = NA,
  coord.fixed = FALSE,
  by.col = TRUE,
  ncols = NULL,
  combine = TRUE,
  group.by.x,
  group.by.y
)
}
\arguments{
\item{object}{seurat object}

\item{features}{gene list to show}

\item{assay}{seurat object assay to draw, defaults to the active assay}

\item{cols}{customized color list}

\item{col.min}{Minimum scaled average expression threshold}

\item{col.max}{Maximum scaled average expression threshold}

\item{dot.min}{set minimum size of dots}

\item{dot.scale}{The fraction of cells at which to draw the smallest dot
(default is 0). All cell groups with less than this expressing the given
gene will have no dot drawn.}

\item{idents}{Identity classes to include in plot (default is all)}

\item{cluster.idents}{Whether to order identities by hierarchical clusters
based on given features, default is FALSE}

\item{scale}{Determine whether the data is scaled, TRUE for default}

\item{legend}{Combine legends into a single legend}

\item{scale.by}{Scale the size of the points by 'size' or by 'radius'}

\item{keep.scale}{How to handle the color and size scale across multiple plots. Same as Seurat::FeaturePlot()}

\item{scale.min}{Set lower limit for scaling, use NA for default}

\item{scale.max}{Set upper limit for scaling, use NA for default}

\item{coord.fixed}{Plot cartesian coordinates with fixed aspect ratio}

\item{by.col}{If splitting by a factor, plot the splits per column with the features as rows}

\item{ncols}{For several features, set number of columns in merged plot}

\item{combine}{Combine plots into a single \code{\link[patchwork]{patchwork}ed}}

\item{group.by.x}{First factor to group the cells by}

\item{group.by.y}{Second factor to group the cells by}
}
\value{
A ggplot object
}
\description{
Drawing DotPlot showing gene expression levels across two classification groups
}
\examples{
ifnb <- LoadData("ifnb")
OrthoDotPlot(ifnb, features = c("CD4", "CD86", "CD96"), group.by.x = "stim", group.by.y = "seurat_annotations", legend = T, keep.scale = 'feature')
OrthoDotPlot(ifnb, features = c("CD4", "CD86", "CD96"), group.by.x = "stim", group.by.y = "seurat_annotations", legend = T, keep.scale = 'all')
}
\seealso{
\code{RColorBrewer::brewer.pal.info}
}
\concept{visualization}
